PHP 8.5 compatibility
Replace non-canonical cast (double)
Patch Guzzle for php 8.5

+++ galette/install/scripts/upgrade-to-1.10.php
@@ -113,7 +113,7 @@ class UpgradeTo110 extends AbstractUpdater
                         if ($paypal[ContributionsTypes::PK] == $k) {
                             $queries[] = array(
                                 'id'   => $k,
-                                'amount' => (double)$paypal['amount']
+                                'amount' => (float)$paypal['amount']
                             );
                         }
                     }
